Move customizations of marshallable elements into MarshalledElementFactory so it's all in one place.
