+++ SharpGen/Transform/MarshalledElementFactory.cs
@@ -1,4 +1,5 @@
-﻿using SharpGen.CppModel;
+﻿using SharpGen.Config;
+using SharpGen.CppModel;
 using SharpGen.Logging;
 using SharpGen.Model;
 using System;
@@ -27,9 +28,8 @@ namespace SharpGen.Transform
         /// </summary>
         /// <typeparam name="T">The C# type to return</typeparam>
         /// <param name="marshallable">The marshallable element to create the C# type from.</param>
-        /// <param name="isTypeUsedInStruct">if set to <c>true</c> this type is used in a struct declaration.</param>
         /// <returns>An instantiated C# type</returns>
-        public T Create<T>(CppMarshallable marshallable, bool isTypeUsedInStruct = false) where T : CsMarshalBase, new()
+        private T CreateCore<T>(CppMarshallable marshallable) where T : CsMarshalBase, new()
         {
             CsTypeBase publicType = null;
             CsTypeBase marshalType = null;
@@ -157,46 +157,26 @@ namespace SharpGen.Transform
             if (csMarshallable.HasPointer)
             {
                 marshalType = typeRegistry.ImportType(typeof(IntPtr));
-                if (isTypeUsedInStruct)
+                
+                if (publicTypeName == "void")
                 {
                     publicType = typeRegistry.ImportType(typeof(IntPtr));
-                }
-                else
-                {
-                    if (publicTypeName == "void")
-                        publicType = typeRegistry.ImportType(typeof(IntPtr));
-
-                    marshalType = typeRegistry.ImportType(typeof(IntPtr));
                 }
 
                 switch (publicTypeName)
                 {
                     case "char":
                         publicType = typeRegistry.ImportType(typeof(string));
-                        marshalType = typeRegistry.ImportType(typeof(IntPtr));
                         break;
                     case "wchar_t":
                         publicType = typeRegistry.ImportType(typeof(string));
-                        marshalType = typeRegistry.ImportType(typeof(IntPtr));
                         csMarshallable.IsWideChar = true;
                         break;
                 }
             }
-            else
-            {
-                if (isTypeUsedInStruct)
-                {
-                    // Special case for Size type, as it is default marshal to IntPtr for method parameter
-                    if (publicType.QualifiedName == globalNamespace.GetTypeName(WellKnownName.PointerSize))
-                        marshalType = null;
-                }
-            }
 
             csMarshallable.PublicType = publicType;
-            csMarshallable.HasMarshalType = (marshalType != null || marshallable.IsArray);
-            if (marshalType == null)
-                marshalType = publicType;
-            csMarshallable.MarshalType = marshalType;
+            csMarshallable.MarshalType = marshalType ?? publicType;
 
             return csMarshallable;
         }
@@ -211,6 +191,160 @@ namespace SharpGen.Transform
                 || marshalType.Type == typeof(ushort)
                 || marshalType.Type == typeof(sbyte)
                 || marshalType.Type == typeof(ulong);
+        }
+
+        public CsMarshalBase Create(CppMarshallable cppMarshallable)
+        {
+            return CreateCore<CsMarshalBase>(cppMarshallable);
+        }
+
+        public CsReturnValue Create(CppReturnValue cppReturnValue)
+        {
+            return CreateCore<CsReturnValue>(cppReturnValue);
+        }
+
+        public CsField Create(CppField cppField)
+        {
+            var field = CreateCore<CsField>(cppField);
+
+            if (field.HasPointer && field.PublicType != typeRegistry.ImportType(typeof(string)))
+            {
+                field.PublicType = typeRegistry.ImportType(typeof(IntPtr));
+            }
+            else if (field.PublicType.QualifiedName == globalNamespace.GetTypeName(WellKnownName.PointerSize))
+            {
+                // Special case for Size type, as it is default marshal to IntPtr for method parameter
+                field.MarshalType = field.PublicType;
+            }
+            return field;
+        }
+
+        public CsParameter Create(CppParameter cppParameter)
+        {
+            var param = CreateCore<CsParameter>(cppParameter);
+
+            var cppAttribute = cppParameter.Attribute;
+            var paramRule = cppParameter.GetMappingRule();
+
+            bool hasArray = cppParameter.IsArray || ((cppAttribute & ParamAttribute.Buffer) != 0);
+            bool hasParams = (cppAttribute & ParamAttribute.Params) == ParamAttribute.Params;
+            bool isOptional = (cppAttribute & ParamAttribute.Optional) != 0;
+            bool hasPointer = param.HasPointer;
+
+            var publicType = param.PublicType;
+            var marshalType = param.MarshalType;
+
+            CsParameterAttribute parameterAttribute = CsParameterAttribute.In;
+
+            if (hasArray)
+            {
+                hasPointer = true;
+            }
+
+            // --------------------------------------------------------------------------------
+            // Pointer - Handle special cases
+            // --------------------------------------------------------------------------------
+            if (hasPointer)
+            {
+                marshalType = typeRegistry.ImportType(typeof(IntPtr));
+
+                // --------------------------------------------------------------------------------
+                // Handling Parameter Interface
+                // --------------------------------------------------------------------------------
+                if (publicType is CsInterface publicInterface)
+                {
+                    // Force Interface** to be ParamAttribute.Out when None
+                    if (cppAttribute == ParamAttribute.In)
+                    {
+                        if (cppParameter.Pointer == "**")
+                            cppAttribute = ParamAttribute.Out;
+                    }
+
+                    if ((cppAttribute & ParamAttribute.In) != 0 || (cppAttribute & ParamAttribute.InOut) != 0)
+                    {
+                        parameterAttribute = CsParameterAttribute.In;
+
+                        // Force all array of interface to support null
+                        if (hasArray)
+                        {
+                            isOptional = true;
+                        }
+                    }
+                    else if ((cppAttribute & ParamAttribute.Out) != 0)
+                        parameterAttribute = CsParameterAttribute.Out;
+                }
+                else
+                {
+                    // If a pointer to array of bool are handle as array of int
+                    if (param.IsBoolToInt && (cppAttribute & ParamAttribute.Buffer) != 0)
+                        publicType = typeRegistry.ImportType(typeof(int));
+
+                    if ((cppAttribute & ParamAttribute.In) != 0)
+                    {
+                        var fundamentalType = (publicType as CsFundamentalType)?.Type;
+                        parameterAttribute = fundamentalType == typeof(IntPtr)
+                                            || publicType.Name == globalNamespace.GetTypeName(WellKnownName.FunctionCallback)
+                                            || fundamentalType == typeof(string)
+                                                    ? CsParameterAttribute.In
+                                                    : CsParameterAttribute.RefIn;
+                    }
+                    else if ((cppAttribute & ParamAttribute.InOut) != 0)
+                    {
+                        if ((cppAttribute & ParamAttribute.Optional) != 0)
+                        {
+                            publicType = typeRegistry.ImportType(typeof(IntPtr));
+                            parameterAttribute = CsParameterAttribute.In;
+                        }
+                        else
+                        {
+                            parameterAttribute = CsParameterAttribute.Ref;
+                        }
+
+                    }
+                    else if ((cppAttribute & ParamAttribute.Out) != 0)
+                        parameterAttribute = CsParameterAttribute.Out;
+
+                    // Handle void* with Buffer attribute
+                    if (cppParameter.GetTypeNameWithMapping() == "void" && (cppAttribute & ParamAttribute.Buffer) != 0)
+                    {
+                        hasArray = false;
+                        parameterAttribute = CsParameterAttribute.In;
+                    }
+                    else if (publicType is CsFundamentalType fundamental && fundamental.Type == typeof(string)
+                        && (cppAttribute & ParamAttribute.Out) != 0)
+                    {
+                        publicType = typeRegistry.ImportType(typeof(IntPtr));
+                        parameterAttribute = CsParameterAttribute.In;
+                        hasArray = false;
+                    }
+                    else if (publicType is CsStruct structType &&
+                                (parameterAttribute == CsParameterAttribute.Out || hasArray || parameterAttribute == CsParameterAttribute.RefIn || parameterAttribute == CsParameterAttribute.Ref))
+                    {
+                        // Set MarshalledToNative on structure to generate proper marshalling
+                        structType.MarshalledToNative = true;
+                    }
+                }
+            }
+            else if (publicType is CsStruct structType && parameterAttribute != CsParameterAttribute.Out)
+            {
+                structType.MarshalledToNative = true;
+            }
+
+            param.HasPointer = hasPointer;
+            param.Attribute = parameterAttribute;
+            param.IsArray = hasArray;
+            param.HasParams = hasParams;
+            param.HasPointer = hasPointer;
+            param.PublicType = publicType ?? throw new ArgumentException("Public type cannot be null");
+            param.MarshalType = marshalType;
+            param.IsOptional = isOptional;
+
+            // Force IsString to be only string (due to Buffer attribute)
+            if (param.IsString)
+            {
+                param.IsArray = false;
+            }
+            return param;
         }
     }
 }
